chore(backup): Rename resource files to snake_case and remove unused DockerManagerResource
Addresses bot review: file naming convention violation (CamelCase → snake_case)
and dead code (DockerManagerResource registered but not consumed by any asset).
